requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-json-widget==2.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
slugify==0.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
